refactor(#1/domain): 응답 관련 패키지 리팩토링
