+++ src/main/java/org/sopt/ticketbay/global/exception/GlobalExceptionHandler.java
@@ -3,10 +3,9 @@ package org.sopt.ticketbay.global.exception;
 import jakarta.servlet.http.HttpServletRequest;
 import java.time.Instant;
 import lombok.extern.slf4j.Slf4j;
-import org.sopt.ticketbay.global.response.ApiResponseBody;
-import org.sopt.ticketbay.global.response.ErrorCode;
-import org.sopt.ticketbay.global.response.ErrorMeta;
-import org.sopt.ticketbay.global.response.CommonErrorCode;
+import org.sopt.ticketbay.global.response.dto.ApiResponseBody;
+import org.sopt.ticketbay.global.response.code.ErrorCode;
+import org.sopt.ticketbay.global.response.dto.ErrorMeta;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
