+++ src/main/java/org/sopt/ticketbay/global/response/code/ErrorCode.java
@@ -1,4 +1,4 @@
-package org.sopt.ticketbay.global.response;
+package org.sopt.ticketbay.global.response.code;
 
 public interface ErrorCode extends ResponseCode {
 }
