+++ src/main/java/org/sopt/ticketbay/global/response/code/ResponseCode.java
@@ -1,4 +1,4 @@
-package org.sopt.ticketbay.global.response;
+package org.sopt.ticketbay.global.response.code;
 
 public interface ResponseCode {
     int getStatus();
