+++ src/main/java/org/sopt/ticketbay/global/response/dto/ErrorMeta.java
@@ -1,4 +1,4 @@
-package org.sopt.ticketbay.global.response;
+package org.sopt.ticketbay.global.response.dto;
 
 import java.time.Instant;
